Updated the map ordinal type to unsigned long ints. This should prevent the number of mesh entities from being an issue.

+++ core/src/TensorField.hpp
@@ -40,7 +40,7 @@ class TensorField
 
     //@{
     //! Typedefs.
-    typedef int                                      OrdinalType;
+    typedef unsigned long int                        OrdinalType;
     typedef Teuchos::Comm<int>                       Communicator_t;
     typedef Teuchos::RCP<const Communicator_t>       RCP_Communicator;
     typedef Teuchos::RCP<Domain>                     RCP_Domain;
